Pet following patch for Nethack 3.4.3
The following YANI was posted recently to RGRN:

[4]: Follow the Leader: when you have multiple pets and you climb
stairs, not only should pets adjacent to you follow you but also pets
adjacent to pets that are adjacent to you (etc.).

This is a simple patch which does that.

Breakdown of files changed:

include/config.h
  add definition of PET_FOLLOW
src/dog.c
  add findfollowers() which finds and marks all pets adjacent to you
  or to another pet which is going to be following you, and then have
  all marked pets follow you in keepdogs()

Web: http://bilious.homelinux.org/?238