Enlightenment dipping patch for Nethack 3.4.3
In rgrn, Displaced Avenger suggested that dipping a loadstone into a
potion of enlightenment should turn it into a flint stone.  Here is a
patch that does that, as well as a couple other similar things.
Cursed bags of holding become uncursed, uncursed bags of holding are
blessed, and very heavy iron balls (generated by being punished while
already punished) are reduced in weight to that of a normal iron ball.

The reasoning behind picking those three is that for all of them,
weight is one of their defining characteristics, and the weight
is magical in nature.  (Yes, I know, debatable in the case of the
iron ball.)  So a potion of enlightenment isn't going to make your
bronze plate mail lighter.

I think it's a fun interaction for a potion that isn't really used
that much.

Breakdown of files changed:

include/config.h
  add definition of ENLIGHTENMENT_DIP

src/potion.c
  add cases for dipping objects into potions of enlightenment
  in dodip()

Web: http://bilious.homelinux.org/?239