Named Monsters v1.0 - Quantum Mechanics and erinyes have names
In NetHack 3.0 the Quantum Mechanics were named after famous real and fictional
quantum physicists. When the way to which named monsters were referred changed
("Peter bites!" instead of "Peter the wolf bites!"), meeting pre-named monsters
got a little confusing. So it was removed.

I've created a patch that keeps a pre-named monster's name unknown until introduced
in some manner (chatting with them or reading their mind). Quantum Mechanics are
randomly named from a list that includes all of the old names plus a couple of new
ones. Not every Quantum Mechanic gets a name, and if they do, it may not be unique.
In order to maintain bonesfile compatibility the names are removed when saving to a bonesfile.

The patch also names the erinyes after the mythological erinyes. They are randomly
ordered, and will not repeat. If an erinys is in a bonesfile, it will be renamed when
that file is loaded (possibly to its old name).

Web: http://bilious.homelinux.org/?181